AESP-Microservice-Project/services/user-service/requirements.txt
# --- Web Framework & Core ---
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0

# --- MySQL driver & Security ---
# Thêm các thư viện bổ trợ để pymysql chạy mượt với MySQL 8.0
pymysql==1.1.0
cryptography==41.0.5

# --- Auth & Security ---
# Flask-JWT-Extended đã bao gồm PyJWT rồi, nhưng để bản này cho chắc chắn
PyJWT==2.8.0
Flask-JWT-Extended==4.5.3

# --- Message Queue (RabbitMQ) ---
# Thư viện pika để "nói chuyện" với RabbitMQ
pika==1.3.2

# --- WSGI Server ---
# Dùng Gunicorn để chạy production trong Docker
gunicorn==21.2.0

# --- Testing ---
pytest==7.4.3